pyproject.toml
[tool.poetry]
name = "test-stripe-backend"
version = "0.1.0"
description = ""
authors = ["zraza-riis <zraza@riis.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.2"
stripe = "^8.5.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
